requirements.txt
typer==0.16.0
langchain==0.3.25
langgraph==0.4.8
langchain-community==0.3.25
langchain-core==0.3.67
langchain-ollama==0.3.3
ollama==0.5.1
langchain-google-genai==2.1.3
langchain-openai==0.3.27
torch>=2.4.0
torchaudio>=2.4.0
transformers>=4.53.0
timm>=1.0.17
diskcache==5.6.1
gradio>=5.46.1
open_clip_torch
ultralytics
laion-clap
opencv-python
flask
rich
python-dotenv
pandas
librosa
huggingface_hub[hf_xet]
opencc-python-reimplemented
